+++ ydb/_apis.py
@@ -12,10 +12,6 @@ if typing.TYPE_CHECKING:
         ydb_topic_v1_pb2_grpc,
     )
 
-    from ._grpc.v4.draft import (
-        ydb_dynamic_config_v1_pb2_grpc,
-    )
-
     from ._grpc.v4.protos import (
         ydb_status_codes_pb2,
         ydb_discovery_pb2,
@@ -26,9 +22,6 @@ if typing.TYPE_CHECKING:
         ydb_common_pb2,
     )
 
-    from ._grpc.v4.draft.protos import (
-        ydb_dynamic_config_pb2,
-    )
 else:
     from ._grpc.common import (
         ydb_cms_v1_pb2_grpc,
@@ -39,10 +32,6 @@ else:
         ydb_topic_v1_pb2_grpc,
     )
 
-    from ._grpc.common.draft import (
-        ydb_dynamic_config_v1_pb2_grpc,
-    )
-
     from ._grpc.common.protos import (
         ydb_status_codes_pb2,
         ydb_discovery_pb2,
@@ -51,10 +40,6 @@ else:
         ydb_value_pb2,
         ydb_operation_pb2,
         ydb_common_pb2,
-    )
-
-    from ._grpc.common.draft.protos import (
-        ydb_dynamic_config_pb2,
     )
 
 
@@ -127,12 +112,3 @@ class TopicService(object):
     DropTopic = "DropTopic"
     StreamRead = "StreamRead"
     StreamWrite = "StreamWrite"
-
-
-class DynamicConfigService(object):
-    Stub = ydb_dynamic_config_v1_pb2_grpc.DynamicConfigServiceStub
-
-    ReplaceConfig = "ReplaceConfig"
-    SetConfig = "SetConfig"
-    GetConfig = "GetConfig"
-    GetNodeLabels = "GetNodeLabels"
